pyproject.toml
[project]
name = "nba-mock-draft"
version = "0.1.0"
description = "2026 NBA Mock Draft Prediction Web Application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "black>=26.1.0",
    "playwright>=1.58.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py311"]
